requirements.txt
AutoRAG[parse, ko]>=0.3.2
click
python-dotenv
